refactor: created day of weeks widgets
